Extract src/settings.ts — the settings window
The tab list, the declarative control table, the dispatcher and the dialog's own
handlers, plus the Usage panel's hover tooltip (parented to <body> so a
renderSettings() rebuild can't drop it mid-hover). 330 lines; main.ts 3,472 ->
3,189.

One deviation from the per-callee hook convention, stated because it is a
deviation: this dialog is a control panel and owns none of what it changes —
theme, sidebar sort, launch engine, terminal font, token scan. Seven separate
setters would have been noise, so they arrive as one SettingsHost object of
no-ops. Same shape as seam rule 2, just gathered.

Two things moved to state.ts rather than becoming hooks: `termFontSize` (a
persisted preference like the rest; applyFontSize stays in main.ts because only
it holds the xterm handles) and `SORT_META`. I first duplicated SORT_META into
the new module and caught it — the rail button and the settings segment must
render the same glyph and description, so it is now defined once beside
SORT_MODES and imported by both.

The moved block diffs identical against HEAD apart from `export`, the host.
prefix on those seven calls, and the two direct writes to termFontSize.

Verified through the app's own ⚙ button rather than a dynamic import: all five
tabs render (tasks reads the preferences that moved to ./tasks last commit, usage
renders ./usageview), and the host dispatch works both ways — picking Light then
Dark sets data-theme on <html>, and picking "Needs you first" flips the rail
glyph from ≡ to ◆.
